add init batching
Former-commit-id: a92c8b71db6673f57514be6fc8cad32115408191

+++ crates/reth-tracing-ext/src/lib.rs
@@ -1,7 +1,6 @@
 use std::{fmt::Debug, path::Path, sync::Arc};
 
 use brontes_types::structured_trace::{TransactionTraceWithLogs, TxTrace};
-use num_cpus::get;
 use reth_beacon_consensus::BeaconConsensus;
 use reth_blockchain_tree::{
     externals::TreeExternals, BlockchainTree, BlockchainTreeConfig, ShareableBlockchainTree,
@@ -64,7 +63,7 @@ pub struct TracingClient {
 }
 
 impl TracingClient {
-    pub fn new(db_path: &Path, handle: Handle, max_tasks: u32) -> (TaskManager, Self) {
+    pub fn new(db_path: &Path, handle: Handle, max_tasks: u64) -> (TaskManager, Self) {
         let task_manager = TaskManager::new(handle);
         let task_executor: reth_tasks::TaskExecutor = task_manager.executor();
 
@@ -124,7 +123,7 @@ impl TracingClient {
             fee_history,
         );
 
-        let tracing_call_guard = BlockingTaskGuard::new(max_tasks);
+        let tracing_call_guard = BlockingTaskGuard::new(max_tasks as u32);
 
         let trace = TraceApi::new(provider, api.clone(), tracing_call_guard);
